Assign setbfree rotary speed select to cc 65
Currently rotary.speed-preset is used to control the rotary effect,
but this is not very flexible as it only allows for stop/slow/fast.

Using rotary.speed-select instead will allow stop/slow/fast to be set
independently on both the horn and the drum.

This change maps rotary.speed-select to controller 65 so that it can
be used in the ui.