Fix top scale point for boolean parameters (#596)
Set the top scale point to 0.005 rather than 1.0 for boolean
parameters such as chorus1.
Otherwise, if a snapshot has been saved with the value
of the parameter set to 0.005, it will be rounded down to 0
when the corresponding parameter page is scrolled to in the
UI, resulting in the parameter suddenly being turned off
without warning.